requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
python-multipart==0.0.6
sqlalchemy[asyncio]==2.0.25
pytest==7.4.4
httpx==0.26.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pydantic-settings==2.1.0 